Set up store and Redux Toolkit and update factories to use the standard factory functions.
