+++ src/store/store.ts
@@ -1,21 +1,34 @@
-import { applyMiddleware, combineReducers, createStore } from "redux";
-import thunk from "redux-thunk";
-import { composeWithDevTools } from "redux-devtools-extension";
+import { TSFixMe } from "@canonical/react-components";
+import { configureStore } from "@reduxjs/toolkit";
 
-import checkAuth from "app/check-auth";
 import generalReducer from "app/general";
+import checkAuth from "store/middleware/check-auth";
 import jujuReducer from "juju/reducer";
 import { modelPollerMiddleware } from "store/middleware/model-poller";
-import uiReducer from "ui";
+import uiReducer from "store/ui";
+import { JujuState, GeneralState } from "../types";
+import { UIState } from "./ui/types";
 
-const reduxStore = createStore(
-  combineReducers({
-    general: generalReducer,
+const store = configureStore({
+  // Order of the middleware is important
+  middleware: (getDefaultMiddleware) => [
+    checkAuth,
+    ...getDefaultMiddleware(),
+    modelPollerMiddleware,
+  ],
+  reducer: {
+    general: generalReducer as () => TSFixMe,
     juju: jujuReducer,
     ui: uiReducer,
-  }),
-  // Order of the middleware is important
-  composeWithDevTools(applyMiddleware(checkAuth, thunk, modelPollerMiddleware))
-);
+  },
+});
 
-export default reduxStore;
+// This can be replaced with the returned store type once each slice has been
+// migrated to TypeScript.
+export type RootState = {
+  general: GeneralState;
+  juju: JujuState;
+  ui: UIState;
+};
+
+export default store;
