+++ src/store/ui/index.ts
@@ -1,0 +1,1 @@
+export { default, actions } from "./slice";
