+++ src/store/ui/selectors.ts
@@ -1,0 +1,22 @@
+import { createSelector } from "@reduxjs/toolkit";
+
+import { RootState } from "store/store";
+
+const slice = (state: RootState) => state.ui;
+
+// ----- Exported functions
+
+export const isUserMenuActive = createSelector(
+  [slice],
+  (sliceState) => sliceState.userMenuActive
+);
+
+export const isConfirmationModalActive = createSelector(
+  [slice],
+  (sliceState) => sliceState.confirmationModalActive
+);
+
+export const isSideNavCollapsed = createSelector(
+  [slice],
+  (sliceState) => sliceState.sideNavCollapsed
+);
