+++ src/store/ui/types.ts
@@ -1,0 +1,5 @@
+export type UIState = {
+  confirmationModalActive: boolean;
+  userMenuActive: boolean;
+  sideNavCollapsed: boolean;
+};
